Add files via upload
The recommendation system of amazon follows the principle of generating product based recommendations which means measuring the similarities between two products and then recommend the most similar product to each user. 
I tried to use fewer Python libraries I can for creating this recommendation system. To work with data I will be using only pandas and NumPy library in Python. The dataset does not have name of product in it, it only had product id so the score of the product reviews becomes the important feature for such kind of datasets.